refactor(api): reorganize hooks and remove deprecated API client
